:recycle: (870): move pipeline task base plumbing into celery infra
