+++ mcr-core/mcr_meeting/app/infrastructure/celery_consumer.py
@@ -1,7 +1,7 @@
 from typing import Any
 
 import celery.app.trace  # type: ignore[import-untyped]
-from celery import Celery
+from celery import Celery, Task
 from celery.signals import setup_logging as celery_setup_logging
 
 from mcr_meeting.app.configs.base import CelerySettings
@@ -54,3 +54,28 @@ celery_worker.conf.broker_transport_options = {
     "unacked_index_key": "unacked_index_transcription",
     "unacked_mutex_key": "unacked_mutex_transcription",
 }
+
+
+class MeetingPipelineTask(Task[Any, Any]):  # type: ignore[explicit-any]
+    def set_task_context(self, meeting_id: int, owner_keycloak_uuid: str) -> None:
+        raise NotImplementedError
+
+    def handle_failure(
+        self, meeting_id: int, owner_keycloak_uuid: str, error_code: str
+    ) -> None:
+        raise NotImplementedError
+
+    def before_start(  # type: ignore[explicit-any]
+        self, task_id: str, args: tuple[Any, ...], kwargs: dict[str, Any]
+    ) -> None:
+        self.set_task_context(args[0], args[1])
+
+    def on_failure(  # type: ignore[explicit-any]
+        self,
+        exc: Exception,
+        task_id: str,
+        args: tuple[Any, ...],
+        kwargs: dict[str, Any],
+        einfo: object,
+    ) -> None:
+        self.handle_failure(args[0], args[1], error_code=self.name)
